Intermediate files - test that temp files properly cleaned + Tests - Allow diy in child

+++ t/TestUtils.pm
@@ -186,25 +186,29 @@ sub with_fork(&&)
 	my $tochild = new IO::Pipe;
 
 	if (my $pid = fork()) {
-		$fromchild->reader();
-		$fromchild->autoflush(1);
-		$fromchild->blocking(1);
-		binmode $fromchild;
-
-		$tochild->writer();
-		$tochild->autoflush(1);
-		$tochild->blocking(1);
-		binmode $tochild;
-
+		my $child_exited = 0;
 		{
+			local $SIG{CHLD} = sub { $child_exited = 1; };
+			$fromchild->reader();
+			$fromchild->autoflush(1);
+			$fromchild->blocking(1);
+			binmode $fromchild;
+
+			$tochild->writer();
+			$tochild->autoflush(1);
+			$tochild->blocking(1);
+			binmode $tochild;
+
 			alarm ALARM_FOR_FORK_TESTS; # protect from hang in case our test fail
-			local $SIG{CHLD} = sub { print STDERR "Unexpected child exit"; die; };
 			$parent_cb->($tochild, $fromchild);
 			alarm 0;
 		}
-		kill 'USR1', $pid; # this PID is exists and it's our child, otherwise we'd exit in SIG CHLD above
-		while(waitpid($pid, 0) != -1){ };
+		unless ($child_exited) {
+			kill 'USR1', $pid; # this PID is exists and it's our child, otherwise we'd exit in SIG CHLD above
+			while(waitpid($pid, 0) != -1){ };
+		}
 	} else {
+		local $SIG{USR1} = sub { exit(1); };
 		$fromchild->writer();
 		$fromchild->autoflush(1);
 		$fromchild->blocking(1);
